+++ packages/mysql/src/mysql.mli
@@ -12,10 +12,6 @@ module Error: sig
   val to_string: t -> string
 
   val serialize: t Serde.Ser.t
-
-  val to_json_string: t -> (string, Serde.error) result
-
-  val to_json: t -> Data.Json.t
 end
 
 module Config: sig