+++ packages/mysql/src/protocol.mli
@@ -114,10 +114,6 @@ module Error: sig
   val to_string: t -> string
 
   val serialize: t Serde.Ser.t
-
-  val to_json_string: t -> (string, Serde.error) result
-
-  val to_json: t -> Data.Json.t
 end
 
 type ok_packet = {